sign in log in
Create a SIGN Up activity with Username and Password. Validation of password should 
happen based on the following rules,Password should contain:
•  uppercase and lower case letters.
•  letters and numbers.
•  special characters.
• Minimum length of the password (the default value is8).
On successful SIGN UP proceed to the next Login activity. Here the user should SIGN 
IN using the Username and Password created during signup activity. If the Username and 
Password are matched then navigate to the next activity which displays a message saying 
“Successful Login” or else display a toast message saying “Login Failed”. The user is 
given only two attempts and after that display a toast message saying “Failed Login 
Attempts” and disable the SIGN IN button. Use Bundle to transfer information from one 
activity to another.