counter
Write a program to create an activity with two buttons START and STOP. On pressing of 
the START button, the activity must start the counter by displaying the numbers from One 
and the counter must keep on counting until the STOP button is pressed. Display the 
counter value in a Text View control.